+++ src/main/java/gift/common/validation/ValidateErrorMessage.java
@@ -1,9 +1,8 @@
-package gift.common.exception;
+package gift.common.validation;
 
 public class ValidateErrorMessage {
     public static final String INVALID_PRODUCT_NAME_NULL = "상품명은 필수 입력값입니다.";
-    public static final String INVALID_PRODUCT_NAME_PATTERN
-            = "상품이름에는 한글, 영어 대소문자, 숫자, ( ), [ ], +, -, &, /, _만 포함 가능합니다.";
+    public static final String INVALID_PRODUCT_NAME_PATTERN = "상품이름에는 한글, 영어 대소문자, 숫자, ( ), [ ], +, -, &, /, _만 포함 가능합니다.";
     public static final String INVALID_PRODUCT_NAME_LENGTH = "상품명은 최대 15자까지 입력 가능합니다.";
     public static final String INVALID_PRODUCT_PRICE_NULL = "가격은 필수 입력값입니다.";
     public static final String INVALID_PRODUCT_PRICE_RANGE = "가격은 1원 이상, 21억원 이하이어야 합니다.";
