+++ src/test/java/gift/controller/dto/request/ProductRequestTest.java
@@ -2,7 +2,7 @@ package gift.controller.dto.request;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import gift.common.exception.ValidateErrorMessage;
+import gift.common.validation.ValidateErrorMessage;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validation;
 import jakarta.validation.Validator;
